backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
axe-playwright-python==0.1.6
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cssselect2==0.8.0
dataclasses-json==0.6.7
fastapi==0.121.2
filetype==1.2.0
fonttools==4.60.1
frozenlist==1.8.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
greenlet==3.1.1
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-google-genai==3.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.46
MarkupSafe==3.0.3
marshmallow==3.26.1
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
pillow==12.0.0
playwright==1.48.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydyf==0.11.0
pyee==12.0.0
pyparsing==3.2.5
pyphen==0.17.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==8.5.0
tinycss2==1.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
weasyprint==62.1
webencodings==0.5.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zopfli==0.4.0
zstandard==0.25.0
# === NEW: Production Dependencies === 
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
upstash-redis>=1.0.0
email-validator>=2.0.0
